+++ src/Module/Catalog/Catalog_soundcloud.php
@@ -103,8 +103,8 @@ class Catalog_soundcloud extends Catalog
      */
     public function install()
     {
-        $collation = AmpConfig::get('database_collation', 'utf8_unicode_ci');
-        $charset   = AmpConfig::get('database_charset', 'utf8');
+        $collation = AmpConfig::get('database_collation', 'utf8mb4_unicode_ci');
+        $charset   = AmpConfig::get('database_charset', 'utf8mb4');
         $engine    = ($charset == 'utf8mb4') ? 'InnoDB' : 'MYISAM';
 
         $sql = "CREATE TABLE `catalog_soundcloud` (`id` INT( 11 ) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY , " . "`userid` VARCHAR( 255 ) COLLATE $collation NOT NULL , " . "`secret` VARCHAR( 255 ) COLLATE $collation NOT NULL , " . "`authtoken` VARCHAR( 255 ) COLLATE $collation NULL , " . "`catalog_id` INT( 11 ) NOT NULL" . ") ENGINE = $engine DEFAULT CHARSET=$charset COLLATE=$collation";